requirements.txt
streamlit
openai
google-generativeai>=0.3.0
pandas>=1.5.0
numpy>=1.21.0